add no clear

+++ python3/kirarafantasia_bot/image_recognition/gacha_result/extract.py
@@ -3,6 +3,7 @@ import glob
 import os
 import shutil
 from . import setting
+import clover.common
 
 BOUND_BOX_XYWH_LIST = setting.BOUND_BOX_XYWH_LIST
 
@@ -11,14 +12,16 @@ if __name__ == '__main__':
 
     parser = argparse.ArgumentParser(description='extract board animal')
     parser.add_argument('img_ts', help='img_ts')
+    parser.add_argument('--no_clear', action='store_true', help="no clear output dir")
     args = parser.parse_args()
     
     fn_list = glob.glob(os.path.join('image_recognition','screen_sample','*','*','{}.png'.format(args.img_ts)))
     assert(len(fn_list)==1)
     fn = fn_list[0]
     
-    shutil.rmtree('output',ignore_errors=True)
-    os.makedirs('output')
+    if not args.no_clear:
+        shutil.rmtree('output',ignore_errors=True)
+    clover.common.makedirs('output')
 
     img = cv2.imread(fn)
